Moves willDisplay delegate func to CategorySelectionViewController+Delegate
