+++ quavi/quavi/Controllers/CategorySelectionViewController.swift
@@ -44,20 +44,6 @@ class CategorySelectionViewController: UIViewController {
         navigationController?.navigationBar.topItem?.title = "Tour Categories"
     }
 }
-//MARK: -- Extension
-extension CategorySelectionViewController: UICollectionViewDelegate{
-    func collectionView(_ collectionView: UICollectionView, willDisplay cell: UICollectionViewCell, forItemAt indexPath: IndexPath) {
-        
-        let transformation = CATransform3DTranslate(CATransform3DIdentity, -500, 0, 0)
-        cell.layer.transform = transformation
-        cell.alpha = 0.0
-
-        UIView.animate(withDuration: 0.5) {
-            cell.alpha = 1.0
-            cell.layer.transform = CATransform3DIdentity
-        }
-    }
-}
 
 extension CategorySelectionViewController: UICollectionViewDataSource{
     func collectionView(_ collectionView: UICollectionView, numberOfItemsInSection section: Int) -> Int {
